Define inf(RR) in NTL namespace
C++ headers are written such that ADL can be used and `isinf(RR)` defined in NTL namespace is found in the <complex> algorithms.
For some C++ compilers if ADL is not used, `isinf` has to be defined before the <complex> header is used which means the users have to be careful when including this header. ADL is much simpler and less hacky.